protocol: add interface ivi_application extension
ivi_application extension creates ivi_surface objects tied
to a given wl_surface with a given id. The given id can be used in a
shell to identify which application is assigned to a wl_surface and
layout the surface wherever the shell wants. ivi_surface objects can
be used to receive status of wl_surface in the scenegraph of the
compositor.

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA at xddp.denso.co.jp>
---
Changes for v2:
   - Rename "error" to "warning" because meaning of "error" in wayland is fatal.

Changes for v3:
   - Move "warning" from ivi_application to ivi_surface.
   - Squash Makefile.
   - Add description to ivi_surface:destroy.
   - Update description of ivi_application:surface_create.

Changes for v4:
   - Remove detail description of server side from ivi_surface::destroy
   - Add clear discripton what client shall do if it encounters warning in ivi_surface
     ::warning.
   - Add decription what happens when client tries to tie a wl_surface to multiple
     ivi_surfaces.

Changes for v5:
   - apply review comments from mailing list

Changes for v6:
   - the same as v5